-FIx type int price

+++ Models/CartItem.cs
@@ -9,7 +9,7 @@ namespace LapTrinhWebBanHang.Models
     {
         public int ProductID { get; set; }  // ID sản phẩm
         public string ProductName { get; set; }  // Tên sản phẩm
-        public decimal Price { get; set; }  // Giá sản phẩm
+        public int Price { get; set; }  // Giá sản phẩm
         public int Quantity { get; set; }  // Số lượng sản phẩm
         public string Size { get; set; }
         public string ImageUrl { get; set; }  // URL hình ảnh sản phẩm
